+++ spec/rubocop/ast/next_node_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require_relative 'wrapped_arguments_node'
+
+RSpec.describe RuboCop::AST::NextNode do
+  it_behaves_like 'wrapped arguments node', 'next'
+end